# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce API",
    "version": "1.0.0",
    "description": "API documentation for the E-Commerce platform"
  },
  "servers": [
    {
      "url": "http://localhost:3300"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/coupons": {
      "post": {
        "summary": "Tambahkan kupon baru",
        "tags": [
          "Coupons"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "discount_percentage"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "example": "DISKON50"
                  },
                  "discount_percentage": {
                    "type": "integer",
                    "example": 50
                  },
                  "valid_from": {
                    "type": "string",
                    "format": "date"
                  },
                  "valid_until": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Ambil semua kupon yang masih aktif (belum soft delete)",
        "tags": [
          "Coupons"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/coupons-publish": {
      "get": {
        "summary": "Ambil semua kupon yang masih aktif, untuk ditampilkan ke user (belum soft delete)",
        "tags": [
          "Coupons"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/coupons/{code}": {
      "get": {
        "summary": "Ambil detail kupon berdasarkan kode",
        "tags": [
          "Coupons"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Kode kupon"
          }
        ]
      }
    },
    "/api/coupons/{id}": {
      "put": {
        "summary": "Update an existing coupon",
        "description": "Update the details of an existing coupon, including code, discount percentage, validity, usage limits, and status.",
        "tags": [
          "Coupons"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The coupon ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "The unique code for the coupon"
                  },
                  "discount_percentage": {
                    "type": "integer",
                    "description": "Discount percentage (1-100)"
                  },
                  "valid_from": {
                    "type": "string",
                    "format": "date",
                    "description": "The start date for coupon validity"
                  },
                  "valid_until": {
                    "type": "string",
                    "format": "date",
                    "description": "The end date for coupon validity"
                  },
                  "usage_limit": {
                    "type": "integer",
                    "description": "The maximum number of uses for this coupon"
                  },
                  "usage_count": {
                    "type": "integer",
                    "description": "The current number of times the coupon has been used"
                  },
                  "status_publish": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ],
                    "description": "Status indicating whether the coupon is available for use"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ],
                    "description": "The internal status of the coupon (active or inactive)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coupon updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Coupon updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid discount percentage"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Coupon not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Coupon not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Hapus kupon (soft delete)",
        "tags": [
          "Coupons"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID kupon"
          }
        ]
      }
    },
    "/api/validate-coupon": {
      "get": {
        "summary": "Validasi kupon",
        "tags": [
          "Coupons"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Kode kupon yang ingin divalidasi"
          }
        ],
        "responses": {
          "200": {
            "description": "Kupon valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "coupon": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "code": {
                          "type": "string",
                          "example": "DISKON50"
                        },
                        "discount_percentage": {
                          "type": "integer",
                          "example": 50
                        },
                        "valid_from": {
                          "type": "string",
                          "format": "date",
                          "example": "2025-04-01"
                        },
                        "valid_until": {
                          "type": "string",
                          "format": "date",
                          "example": "2025-04-30"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parameter code tidak diberikan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Kode kupon diperlukan"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Kupon tidak ditemukan atau sudah tidak aktif",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Kupon tidak ditemukan atau sudah tidak aktif"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}